+++ main.go
@@ -36,7 +36,7 @@ var (
 )
 
 var (
-	version = "1.0.8"
+	version = "1.0.9"
 )
 
 func main() {